Merges claude/meshql-patterns-skill-u8b7yl. Conflicts in four Cargo.toml
files and Cargo.lock resolved in favor of main's GitHub git deps
(workspace.dependencies pinned to tags): the branch's switch to
crates.io registry deps is superseded — these crates are deliberately
not distributed via crates.io, and the branch's merkql 0.1.1 pin for
meshql-merksql was a downgrade forced by the stale published merksql.

+++ meshql-sqlite/tests/cross_service_cert.rs
@@ -21,8 +21,8 @@ type Coop {
     name: String
 }
 type Query {
-    getFarm(id: ID, at: Int): Farm
-    getFarms(name: String, at: Int): [Farm]
+    getFarm(id: ID, at: Float): Farm
+    getFarms(name: String, at: Float): [Farm]
 }
 "#;
 
@@ -36,9 +36,9 @@ type Coop {
 }
 type Farm { id: ID name: String }
 type Query {
-    getCoop(id: ID, at: Int): Coop
-    getCoops(name: String, at: Int): [Coop]
-    getCoopsByFarm(id: ID, at: Int): [Coop]
+    getCoop(id: ID, at: Float): Coop
+    getCoops(name: String, at: Float): [Coop]
+    getCoopsByFarm(id: ID, at: Float): [Coop]
 }
 "#;
